+++ inst/quarto/build-template-book.R
@@ -1,0 +1,7 @@
+
+# render quarto book ------------------------------------------------------
+
+if(dir.exists(here::here('book'))){
+  system2('quarto', 'render book')
+}
+