+++ inst/quarto/build-template-website.R
@@ -1,0 +1,7 @@
+
+# render quarto website ---------------------------------------------------
+
+if(dir.exists(here::here('website'))){
+  system2('quarto', 'render website')
+}
+