+++ inst/shiny/idshinyapp/global.R
@@ -1,0 +1,10 @@
+library(shiny)
+library(bslib) #supplies the most flexible dashboard options and bootstrap 5 support
+#also works with basic shiny pages but not shinydashboard or bs4Dash
+library(ggplot2)
+library(plotly)
+library(thematic)
+library(idstyle)
+
+thematic_shiny()
+