+++ NEWS.md
@@ -1,3 +1,30 @@
+# idstyle 0.0.1.0009
+
+Lots of updates in this dev version for review:
+
+## Templates available in `inst` folder
+
+- git
+  - .gitattributes & .gitignore
+- quarto
+  - idreport template using quarto extensions for html, pdf, and docx outputs
+- rmarkdown
+  - updated the docx rmd output to use a template docx files
+- r packages
+  - NEWS & README templates
+- rstudio
+  - project templates for r packages & analysis projects
+  - addins for logo text
+- shiny
+  - templates for shiny application (global, ui, server), function to run application, and README
+  
+## Functions & Data in `R` folder
+
+- colors & fonts
+- bootstrap theme
+- other functions/data to support templates described above accessed through R Studio wizard and add-ins.
+
+
 # idstyle 0.0.1
 
 * Initial creation